CON-412 Upgrade to Kotlin 1.5
Updated apiVesion and languageVersion.

Fixed a build issue by including the classgraph-\*.jar dependency in conclave-host-\*.jar.

**Summary of the issue found:**
The task cpp:generateHostSharedJniHeader started to fail after updating the properties apiVersion, languageVersion to 1.5 in the common.gradle file. The task cpp:generateHostSharedJniHeaders depends on the conclave-host-\*.jar and it was failing because the class io.github.classgraph.Resource could not be found. To fix the issue, the path to  io.github.classgraph-*.jar is provided to the command javah.