CON-482 Provide startup and shutdown hooks for the Enclave
Added a state manager to the enclave. The NEW state indicates the enclave was created but not initialised, the STARTED state indicates the enclave is initialised, the CLOSE state indicates the enclave has been closed. 

From now on, the function onStatup is called only once even if the enclave received multiple onOpen requests. The enclave transitions from the NEW state to STARTED after the function onOpen is called.

From now on, the function onClose is called only once even if the enclave receives multiple onClose requests. The enclave transitions from the STARTED state to CLOSED after the function onClose is called. 

From now on, the enclave throws an exception if the host tries to communicate with the enclave and the enclave has not been initialised or the enclave has already been closed.

From now on, the enclave throws an exception if it receives mail but the enclave has not been initialised or it has already been closed.

Added two unit-tests to check that the enclave-host throws an exception if deliverMail and callEnclave are called after the enclave-host is closed.