[CON-425] Intel sunsets PCS v1 and v2 - must check DCAP code
# Changes

Upgraded the azure plugin from version 1.6 to 1.10 (which uses v3 of the DCAP protocol by default and should also support ice lake). Also made some changes to `azure-plugin/download.sh` to make version selection configurable.

# Notes

The plugin comes in flavours for Ubuntu 18.04 and 20.04. At the moment we default to 18.04, which doesn't seem to cause any issues (even on 20.04 systems). Even so we should probably select the correct version in the future. To do this though, we will need a mechanism for downloading at build time (which we currently don't have).

Alternatively, we could make this the responsibility of the user/systems administrator by adding the installation of the plugin (which is distributed by the authors as a .deb) to the machine setup instructions.

Another option might be to replicate the functionality of the plugin (which seems to be fairly simple) in our own code by interacting with Intels REST endpoints ourselves. Though this approach has some pros and cons:

Pros:
- Eliminate a dependency
- More control, e.g. we can select the caching server (rather than always using the azure one)
    - Might make sense to expose some of these settings to the user
    - It may be that the azure plugin _can_ actually do this, I haven't checked!

Cons:
- More code to maintain 
- Potential security problems if we do a bad job

# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?