[CON-624] Add encrypted space in Mail for the sealed state ID and other internal Conclave use
# Changes

Added a field to mail referred to as the "private header", which is encrypted and not available via the public API. It is transmitted after the handshake but before the body of the mail.

# Notes

Implementation was complicated by the use of lazy handshakes in the MailEncryptingStream and MailDecryptingStream classes. In the case of the MES the code has been changed to use a greedy approach where the handshake takes place in the constructor. In the case of the MDS however, this was not possible due to the requirement that it be possible to supply the key after reading the mail header. If this code could also be refactored to use a greedy approach, this would simplify the logic and make the code much easier to follow.

# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?