Reduce the amount of copying done when sealing
The use of OpaqueBytes in the PlaintextAndEnvelope class means there is unnecessary copying of byte arrays that occurs. OpaqueBytes is an immutable class part of the public API and does the copying to prevent inadvertent modification of the underlying byte array. The enclave's sealing functionality is internal only and so doesn't need to use OpaqueBytes at all.