CON-482 Provide startup and shutdown hooks for the Enclave
Fixed two concurrency issues. The function onOpen and onClose must be synchronised with receiveFromUntrustedHost and receiveMail function.

Summary of issues:
The first issue that required to be fixed was the fact that onOpen and onClose were not being synchronised with receiveFromUntrustedHost and receiveMail because the  onOpen and onClose function were not locking the same object as receiveFromUntrustedHost and receiveMail. onOpen and onClose were locking an AdminHandler object while the receive* functions were locking the enclave object.

The second issue that needed addressing was related to a multi-thread enclave. When on a multi-thread enclave, it would be possible to execute onClose while another thread was still executing a receive* function. 

Solution applied:
For the first issue, it was necessary to guaranty that the same lock is used by all four functions that required to be in sync.
For the second issue, it was necessary to use a conditional variable and a counter to guaranty that the onClose function waits until all receive* execution finish.