CON-627: Enclave persistent transactional K-V store
Enclave has a new property, `persistentMap`, which is a java.util.Map mapping string keys to arbitrary byte arrays. This map is designed to be persisted by the host such that its data survives enclave restarts and is protected from roll back attacks by the host.

The contents of the Map are stored in a “sealed state” blob which is created at the end of each `receiveFromUntrustedHost` and `receiveMail` call via a new `StoreSealedState` command. The blob is encrypted using the MRSIGNER key. The most recent blob must be passed into `EnclaveHost.start` on restart. To make sure the host does this, and doesn’t use an older copy, the enclave tags each outbound mail created from each receive call with the unique ID for that state and also the state ID that particular client last saw. The client `PostOffice` checks that the “last seen state ID” attached to the mail is correct. This way if the host rolls back the state then as long as the client receives further mail it will detect the roll back.

With the introduction of a secure persistent store, there’s no need for the “mail-to-self” pattern. Therefore, mail redelivery and its associated features, such as mail acknowledgement, are no longer required and have been deleted. This includes the `AcknowledgeMail` and `AcknowledgementReceipt` commands and the `acknowledgeMail` method. Writing enclaves should now be easier now that developers don’t have to worry about mail redelivery and using mail-to-self for persistence.

In fact the enclave must now guard against mail redelivery in case the host can manipulate enclave’s state by replaying the same mail. This has been achieved by switching the enclave’s encryption to a random session key that’s generated on each enclave start. This automatically prevents old mail from being read by the enclave. This is also consistent with Mail’s new communication-only scope. A random session key also has the benefit of preventing the host from performing attacks by spinning up multiple instances of the same enclave.

The persistent Map is only available to single-threaded enclaves (i.e. ones whose `threadSafe` method returns false). Thread-safe enclaves cannot use the Map and thus they don’t emit a sealed state blob as there’s no state to persist. Support may be added in the future.