CON-627: Removed the batch sequence from the mail private header
The batch sequence number was an attempt at detecting reordering and dropped mail if the enclave sent multiple mail from the same `receiveMail`/`receiveFromUntrustedHost` invocation. However, this is not a concern of rollback detection on the enclave's state and so has been removed. Client detection of reordered/dropped/replayed mail is a separate issue.