CON-672 Building an enclave using Gradle >= 7.0 results in a non-obvius error message
**Summary of the Issue**

Before this commit, the error message "Entry META-INF/MANIFEST.MF is a duplicate but no duplicate handling strategy has been set..." would be displayed when building the hello-world sample using Gradle 7.0.

Gradle 7.0 changed the default way it treats duplicate files in copy task. Now, it is required to explicitly specify the duplicates strategy.

**Solution**
This commit fixes the issue by explicitly setting the duplicate strategy to DuplicatesStrategy.EXCLUDE.
Gradle 7.0 is still not officially supported even after this commit because more testing is necessary but there is not enough time. The docs were updated to let people know that they should use Gradle v6.6.1 or an older version. And that attempting to use a more recent version might lead to unexpected build issues.