CON-703: Fix for pthread_attr_getstack() to return stack address.
# Changes
This fixes the random stack error that is sometimes seen when starting enclaves. The JDK calls pthread_attr_getstack() on the native layer to retrieve the start and length of the stack. Our implementation was returning the length but the start was left as an uninitialised variable. This caused the resulting check inside the JDK to sometimes fail.

# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?
- [ ] If the gradle plugin has been updated, have the gradle plugin tests been run in TeamCity?