CON-698 - Docs - Document the fact that people must unset the environment variable SGX_AESM_ADDR before running Conclave in debug or release mode
Updated the docs to alert people that they must  unset the environment variable SGX_AESM_ADDR before running Conclave in debug or release mode.
Added a check to the EnclaveHost which throws an exception if Conclave is running in debug or release mode and the environment variable SGX_AESM_ADDR is set. A warning message is displayed for the other modes if SGX_AESM_ADDR is set.