CON-729: Introduce single zero header in Mail to support future ConclaveJS
`MailProtocol.SENDER_KEY_TRANSMITTED_V2` was introduced for 1.2 to add support for an internal encrypted space inside Mail in a backwards compatible way. Use this to also add the single zero payload required in the header to support ConclaveJS in browsers. This will avoid us having to create a _V3 protocol for 1.3.