Fix enclave aborts with persistent filesystem
This PR contains two fixes: 

# Fix call to mkfs to allow any supported FAT size
Previously if the user specified a file system size less than 32Mb then the creation of the filesystem would fail. This is because the filesystem type was hardcoded to FAT32 instead of letting FatFS select the FAT type.

Also there was a bug in this code where the options structure was not initialised. This could cause some randomness in the initialisation parameters.

# Return an error instead of throwing an exception when files are not found.
The code that determined which filesystem contained a path was throwing an exception if the filesystem could not be determined. Now instead this will return null. Any function that takes a filename as a parameter that uses this will now return an error and set `errno=ENOENT`. This prevents the enclave from aborting for example when non-essential files are queried (such as `/etc/timezone` as observed in the KDS).