CON-714 EnclaveHost.load does not work in mock mode
Added a utility class called UtilsOS.
Updated the function checkPlatformEnclaveSupport to return early in case the mode is MOCK.
NativeShared.kt checks if the platform is Linux before loading libjvm_host_shared.so. The library is not loaded if the platform is not Linux.  This will allow Conclave mock mode to run natively in non-unix platforms.