CON-617: New EnclaveClient API (#871)
The introduction of enclave session keys means enclave restarts are no longer transparent. Clients now have to deal with possibility the mail they send may not decrypt due to the enclave’s encryption key changing. In addition, the introduction of the rollback resistant persistent map has meant clients need to keep track of the enclave’s state to detect any rollbacks. These features have increased the complexity on the client.

To address these issues, `EnclaveClient` has been introduced. It provides a simple interface for the user to send mail to the enclave, automatically re-trying if it detects the enclave’s key has changed. It also keeps track of the enclave’s state and signals to the user if it detects the enclave’s state has been rolled back.

`EnclaveTransport` has also been introduced to abstract away, from `EnclaveClient`, the details of how mail is transported to and from the host. It defines a protocol for how the host is to behave for an `EnclaveClient` can connect to it. The conclave-web-host and its REST API has been modified to follow this protocol and a `WebEnclaveTransport` class is provided as the client-side implementation. The hello world sample has been updated to use this.