Setting product ID if enclave is configured to use own code signer for KDS and other improvements
`PolicyConstaint.ownCodeSigner` has been changed into `ownCodeSignerAndProductID` to satisfy requirement that a product ID also be provided if code signer is used in policy constraint. This is also a safe default for now.

Converted the `masterKeyType` parameter to be an enum rather than string for better developer experience. Introducing new key types will require new SDK releases, but this is fine as the application enclave would have be re-compiled anyway if strings were used.

Some of the KDS classes have been moved from conclave-common into conclave-enclave as they're only needed there.

Also, improved the checks made by the enclave, including making sure that it's given a KDS private key if it's configured to use a KDS and in release mode.