CON-754 Conclave plugin should set sourceCompatibility, targetCompatiblity, and jvmTarget based on the JDK version being used instead of defaulting to Java 11
**Issue**
Before these changes, the enclave plugin would always set sourceCompatibility, targetCompatibility, and jvmTarget to Java 11. And the user would have to set the sourceCompatibility, targetCompatibility, and jvmTarget inside the ./enclave/build.gradle file if they planned to overwrite the values of those properties. This behaviour causes issues to existing projects that rely on JDK 8 such as Q.A.'s project pqa-con-app-no-op. The test "the attestation after startup" from AttestationTests.kt file was failing when the enclave was being loaded while using JDK 8 and mock mode. 

**Decision**
To improve the developer experience, Conclave's plugin will set the Java version for sourceCompatibility, targetCompatibility, and jvmTarget based on the JDK version. 

**Summary of changes**
Updated the logic used by the enclave plugin to update the sourceCompatibility, targetCompatibility, and jvmTarget. From now on, the properties will be set to Java 11 if the JDK version is 11 or newer. Otherwise, the values are set to Java 8. A info message is logged if sourceCompatibility, targetCompatibility, or jvmTarget are updated by the plugin. 

If the properties are set to Java 8, a info message is logged letting developers know that the recommended version is Java 11