CON-770: Replace backward with forward strings in app-resources-config.json on Windows
# Changes
When building the `conclave-kms` repo, we get this error
```
S C:\Users\RamjiKerai\workspace\conclave-kms> ./gradlew :kms-host:assemble -PenclaveMode=simulation
Starting a Gradle Daemon, 1 incompatible and 3 stopped Daemons could not be reused, use --status for details

> Task :kms-enclave:buildUnsignedGraalEnclaveSimulation FAILED
Warning: Ignoring server-mode native-image argument --no-server.
Fatal error:java.util.regex.PatternSyntaxException: \k is not followed by '<' for named capturing group near index 8
com\r3\keyservice\enclave\enclave.properties
        ^
        at java.base/java.util.regex.Pattern.error(Pattern.java:2027)
        at java.base/java.util.regex.Pattern.escape(Pattern.java:2545)
        at java.base/java.util.regex.Pattern.atom(Pattern.java:2285)
        at java.base/java.util.regex.Pattern.sequence(Pattern.java:2209)
        at java.base/java.util.regex.Pattern.expr(Pattern.java:2068)
        at java.base/java.util.regex.Pattern.compile(Pattern.java:1782)
        at java.base/java.util.regex.Pattern.<init>(Pattern.java:1428)
```
This is due to an issue with Windows paths vs Linux paths.

Specifically when we generate the file `app-resource-config.json` we add paths into it. These paths contain backward slashes as the related Gradle that generates the file is run on Windows.

When we run `native-image` we do it inside a Docker Linux container and that file is read as if we were on a Linux OS, hence causing the pattern matching failure. The fix consists in replacing the backward slashes with forward slashes and this is what we do here.

To investigate, we needed to go into the docker container and see the file that got generated, @ac101m helped on this


# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?
- [ ] If the gradle plugin has been updated, have the gradle plugin tests been run in TeamCity?