CON-763: Bug fix - unable to get KDS key if using useOwnCodeHash() or useOwnCodeSignerAndProductID()
This bug only manifests itself if the enclave returns a new `KDSConfiguration` in `getKdsConfig()`. The enclave now caches the generated policy constraint and uses that when checking the KDS response.

Most of the changes are enabling testing of KDS use in the integration tests. Some of the persistent map and persistence file system tests have been parameterised to test both an MRSIGNER persistence key and a KDS one.