CON-782: Move check related to persistentMap and threadSafe parameter
# Changes
We have an issue with the message returned to the user. This is due to an exception that is thrown at the wrong place, i.e. when the handlers for the enclave are not initialized yet. The fix is to move that at the right place, so to get a correct message./
# Reviewer Checklist

- [ ] Are you happy with the code changes?
- [ ] Is there adequate unit test coverage?
- [ ] Has the external docsite been updated?
- [ ] Have the API docs (KDocs) been updated?
- [ ] Has the internal docsite been updated?
- [ ] Has an item been added to the [Release notes](https://github.com/corda/sgxjvm/blob/master/docs/docs/index.md)?
- [ ] Have all dependencies introduced by this PR been approved?
- [ ] Has licences/3rd-party.md been updated to reflect any new dependencies?
- [ ] If the gradle plugin has been updated, have the gradle plugin tests been run in TeamCity?