CON-800 SpringBoot logging: host is full of debug messages
The shadowJar did not seem to be configured correctly for generating a fat jar for a spring application. This misconfiguration  was causing runtime issues for the host present in the hello-world sample, and the host present in the projects generated by conclave-init.

It was decided that the bootJar task, which is provided by the Spring plugin, should replace the shadowJar task. This strategy should avoid future issues related with the configuration.

Updated the docs.