CON-838: Java 17 support
The previous messaging around Conclave forcing the Java version on the enclave module to be Java 16 (previously 11) was incorrect. Since Gradle automatically configures any project to use the same Java version as the JDK, there's no need for the plugin to do anything in this regard. To that end, that bit of the plugin code has been removed. The messaging around Java compatibility and support in the docs has been simplified.

Java 17 has been set as the default for the hello world sample and the generated project from Conclave Init. Conclave Init itself also now requires Java 17 to run. The integration tests also use Java 17. The rest of the SDK codebase still uses Java 8 to ensure are minimum required Java version is 8.

The Gradle toolchain feature is now used, since it's simpler to configure and also automatically configures the Kotlin JVM target.