CON-857 Removed Gradle warnings related to the tasks setupLinuxExecEnvironment, and untarGraalVM
After Gradle was upgraded to version 7, Gradle started to log warning messages and disabling execution optimizations for the tasks setupLinuxExecEnvironment, untarGraalVM, and signedEnclave*Jar. This commit addresses the warning messages except one which will be fixed later on.

**Details about the issue:**
There is no dependency between the tasks reported by Gradle. The messages are in fact false positives. Gradle thinks there are dependencies between the  tasks because “copyEnclaveCommonHeaders”, “copySgxTools”, and “copySubstrateDependenciesSimulation” are targeting the folder “conclave”. And both “setupLinuxExecEnvironment”, and “untarGraalVM” target as an input a file that is contained by the folder “conclave” but the file is NOT created by “copyEnclaveCommonHeaders”, “copySgxTools”, and “copySubstrateDependenciesSimulation”. 

It is possible to validate the previous argument by simply update  the tasks “copyEnclaveCommonHeaders”, “copySgxTools”, and “copySubstrateDependenciesSimulation” to target a different output folder such as “conclave1”. The messages will not appear after updating the output folder name.

The warning message related to signedEnclave*Jar was caused by Java lambdas which should not be used in Gradle (refer to https://docs.gradle.org/current/userguide/validation_problems.html#implementation_unknown).