CON-950: Fix inability to deliver consecutive mail encrypted with KDS key
The problem stems from the usage of a ThreadLocal to capture the KDS response whilst the enclave is trying to decrypt the inbound mail. That ThreadLocal has been removed completely. Instead the EnclaveHost checks if the mail about to be delivered is encrypted using a KDS key. If it is then it will make the HTTP KDS request and only when the response is received does it deliver the mail bytes to the enclave, along with the private key response. This also removes the need to do an OCALL-ECALL cycle. To avoid unnecessarily making KDS requests for keys the enclave has already cached, the host also keeps track of key specs have been specified and only makes a new private key request if a new key spec is encountered.

The ability for mail to be encrypted with KDS key, alongside the existing enclave session key, means that it's now possible for the user to mix up both KDS and non-KDS encryption for the same (client key, topic) pair. The enclave caches response PostOffice objects based on client key and topic and so if the user mixes up KDS and non-KDS any mail response back to the client might not decrypt. A warning message was added to the enclave if it detects this.

The key name for the persistence KDS key has been changed to something more relevant ("EnclavePersistence"). We can change this now before the KDS goes into production.