CON-960: User specified persistence policy constraint is used as is in the KDS request
This is important as the policy constraint string is used in the key derivation algorithm. Any modification which doesn't impact the actual `EnclaveConstraint` output will still result in a different key (this includes extra whitespace). The tests and documentation have been updated to precisely define how the `useOwnCodeHash` and `useOwnCodeSignerAndProductID` properties modify the user provided portion of the constraint string.