CON-959 Enclave should not check sequence numbers for KDS encrypted Mail
**Summary**
From now on, the Enclave does not check the sequence numbers for KDS encrypted mail. A unit test was added to cover this case.

**In-depth Explanation**
Checking of sequence numbers by the enclave if the mail was encrypted with a KDS key doesn’t make sense. This is because such a Mail can be processed by any number of enclave instances. For example, horizontal scaling of an enclave application. In such a scenario the first mail in the sequence might go to enclave 1 and the second mail to enclave 2. Enclave 2 will then complain that the sequence number hasn’t started from zero.