CON-967: Enclave no longer caches EnclavePostOffice instances when KDS is involved
Just as it doesn't make sense to check for the sequence numbers on inbound mail which are encrypted with a KDS key, it also doesn't make sense to implement increasing sequence numbers for outbound mail encrypted with a KDS key. Thus `postOffice(EnclaveMail)` doesn't return a cached `EnclavePostOffice` if a KDS key is used. This also fixes the design issue that different KDS keys (or non-KDS and KDS keys) couldn't be used on the same (sender, topic) pair, and so the warning message is no longer necessary.

The mock KDS code has been cleaned up and fixed as it was missing some checks and wasn't implementing a correct key derivation process.