CON-885: Fix filesystem issue in deleting directories
# Changes
In the context of the usage of FatFs filesystem, a wrong `NoSuchFileException` was produced when running this code in the Enclave
```
val path = Path.of("/creatingDirs/test1/test2/test3/test4/")
createDirectories(path)

walk(path)
    .sorted(Comparator.reverseOrder())
    .map(Path::toFile)
    .forEach(File::delete)
```
This was due to a wrong handling of the result of the function `f_readdir`, specifically we were returning an error when the directory we were reading from was empty instead of returning a `nullptr` but without setting `errno`.
After fixing this issue, we have also uncovered another one: in `closedir` FatFs functions, we weren't deleting the directories from the container maps. These directories were considered as opened twice.

We fixed the issues by removing the directories from the maps when we close them and by correctly handling the return code from the function `f_readdir`.

I have also addressed some typo in the test classes.