CON-1006: Exceptions thrown during enclave startup are exposed, even in release mode
This is to help troubleshooting startup issues in production. To achieve this, `EnclaveStartException` has been introduced which the enclave will let through, even in release mode. Any exception thrown during the enclave's "open" call will be wrapped in a `EnclaveStartException`. This includes the initialisation of the enclave's file system and user's `onStartup` call.

Also, the serialisation of the exceptions has been improved to include suppressed exceptions.