Optimised the gradle enclave plugin tests
The majority of the enclave plugin tests are now integration tests. This fits better with how they run, which is they require the SDK, which is what the integration tests are. It also allows the plugin tests to be parallelised across the enclave modes, rather than running them as serialised parameterised tests. Further, the tests themselves have been optimised to minimise the number of times the test enclave project is built. All this means the tests can be run as part of each PR, rather than as a very slow path when master or the release branches are built.