containers/conclave-build no longer a Gradle module
It was only a Gradle module to allow the Dockerfile to be published as a separate artifact, to be then picked up by the Gradle plugin. It's simpler to just copy the Dockerfile at build time so that the plugin code has direct access to it. This will also simplify the recent Docker build changes made to master.