CON-1043 Teamcity - Improve how the docker container images are used by each task for the Graal project
These changes modify the way the docker image is used by TeamCity. Before, the docker container image was being passed between TeamCity jobs as compressed a file. Now, the docker container image is pulled in the conventional way, i.e., by using the docker command. TeamCity will always check if the docker container image exist, if it does not, TeamCity will generate and publish a new one to Artifactory.

Please note that:

The docker container image is tagged using a hash.
The hash generated is based on the contents inside the containers directory.
Currently, the hash generated only takes into account files that are tracked by git. Please ensure that is always the case otherwise TeamCity might generate more docker container images than needed.
Each docker container image has a commit_id tag. This tag is useful for knowing what the state of the code was when the image was created. This can be helpful to reproduce issues and debug issues.
The namespace com.r3.sgx was renamed to com.r3.conclave.