CON-1055: Decomissioned the SDK zip
As part of open-sourcing the SDK, the distribution of the SDK zip artifact no longer makes sense and has been removed. The SDK artifacts will be available via public repositories such as Maven Central. An internal SDK repo is still produced and used by the integration tests to ensure the correct build is tested.

Conclave projects generated by the Conclave Init tool no longer have a separate `repo` entry in their Gradle config pointing to a repo directory, and thus the `conclaveRepo` property is no longer needed. The only property that needs to be defined is `conclaveVersion`. Conclave Init now automatically populates this with the current version. This all means the need for a user global gradle.properties file goes away, and Conclave Init now no longer relies on it.

The SDK zip hosted the Conclave Init binary. This is now published alongside the Conclave libraries and the docsite has been updated with the download instructions.