CON-1099: Update docs on support for Apple Silicon
Added a known issue on the known-issues.md page about Conclave not supporting the new Macs with Apple silicon. Added a reference to the known issue on the machine-setup.md page.

I have NOT added this warning on other pages where we mention macOS as it will disrupt the flow of those pages. These pages are:

- enclave-configuration
- enclave-modes
- faq
- mock mode
- release notes
- running hello world
- writing hello world