CON-1112: Fix issue with CONCLAVE_ variable
As pointed out in the PR https://github.com/R3Conclave/conclave-sdk/pull/1171#discussion_r941715652, we have noticed that in case we do not have any CONCLAVE_ variables, we have a bug that prevents to build the container. As this is a common case that might happen, we need a fix.