CON-1138: Make links readable on the docs site
Currently, links are not distinguishable from standard text on the docs site. Only when a user hovers over the link text the user knows it's a link.

I have changed the color of links by adding a CSS variable '--md-typeset-a-color' in the r3-branding.css file.  I have used a lighter version of teal, which wouldn't stand out too much while being readable. The hex color code is #006c94. Needs to test how it looks when published.