CON-1152: Improve visibility of cloud deployment doc
Previously, the first doc under the 'Operations' section described more about non-cloud deployment. The Azure deployment doc was given less priority. However, most users are expected to use the easier and more efficient Azure deployment method. 

With this change, the first doc in the 'Operations' section describes Azure deployment. Non-cloud deployment and IAS are covered in later docs.

As I had to change the file names of the docs, I also had to edit other docs which referenced these file names. Also, changed the mkdocs.yml file to reorder and rename how the docs are displayed on the docs site.

Also, edited all three docs as per the findings in the content audit.