CON-1161 - Integrated gramine Conclave and refactored the code
It is now possible to run Conclave using Gramine but without sgx. This is only experimental and more changes will be added in future commits. To use gramine, Conclave must be built in mock mode and the environment variable CONCLAVE_GRAMINE_ENABLE must be set.

The class EnclaveScanner was created to group together all the functions responsible for searching for enclaves.