CON-1170: Review API links
When a user clicks on some API links on the docs site, the pages don't render correctly. I investigated all the API links on our docs site and fixed the ones which were not displaying correctly—also fixed a few typos that caught my eye. Note that this is not a thorough rewrite of the docs that are changed by this PR.