CON-806 Review INSECURE, STALE, and SECURE enclave constraint options
Updated how the security level is assess. If the security level is set to INSECURE than a client can only communicate with an enclave that is running as INSECURE. If the security level is set to STALE then the client can communicate with a STALE or SECURE enclave. If the security level is set to SECURE then the client can only communicate with a SECURE enclave.