CON-1160: Replace Gramine env variable with Gradle config flag
This is a first step in the direction of integrating Gramine into Conclave. Here we add the related flag in the `build.gradle` files of the enclaves and we allow the `EnclaveHost` to start Gramine enclaves, albeit with some assumptions/dummy behaviour.
The changes should have no impacts on the non-Gramine code.

In order to test a Gramine based enclave (only with a bash script now), I have run a "Hello World" example and checked that a process was created on my Linux machine (in addition to the temp directories).